+++ docs/conf.py
@@ -12,11 +12,11 @@ extensions = [
 ]
 
 intersphinx_mapping = {
-    'python': ('http://python.readthedocs.org/en/latest/', None),
+    'python': ('http://python.readthedocs.io/en/latest/', None),
     'flask': ('http://flask.pocoo.org/docs/latest/', None),
     'flask-sqlalchemy': ('http://flask-sqlalchemy.pocoo.org/latest/', None),
-    'marshmallow': ('http://marshmallow.readthedocs.org/en/latest/', None),
-    'marshmallow-sqlalchemy': ('http://marshmallow-sqlalchemy.readthedocs.org/en/latest/', None),
+    'marshmallow': ('http://marshmallow.readthedocs.io/en/latest/', None),
+    'marshmallow-sqlalchemy': ('http://marshmallow-sqlalchemy.readthedocs.io/en/latest/', None),
 }
 
 primary_domain = 'py'
@@ -34,7 +34,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Flask-Marshmallow'
-copyright = u'2014-2015'
+copyright = u'2014-2016'
 
 version = release = flask_marshmallow.__version__
 exclude_patterns = ['_build']